feat!: moved v1 api under src/v1
BREAKING CHANGE: please update your namespaces if you use any
of the classes directly or composer doesn't like the namespace
change.
